refactor: 폴더 구조 컨벤션 맞추기
