+++ src/components/button/Button.tsx
@@ -1,18 +1,16 @@
 import { PropsWithChildren } from 'react';
 
-import * as S from './index.style';
+import * as S from './Button.style';
 
 export interface ButtonProps extends PropsWithChildren {
     isValid: boolean;
     onClick: () => void;
 }
 
-function Button({ isValid, children, onClick }: ButtonProps) {
+export function Button({ isValid, children, onClick }: ButtonProps) {
     return (
         <S.Button isValid={isValid} onClick={onClick}>
             {children}
         </S.Button>
     );
 }
-
-export default Button;
