+++ src/features/location/current-location-search/index.ts
@@ -1,0 +1,1 @@
+export { CurrentLocationSearch } from './CurrentLocationSearch';
