+++ src/features/location/location-select-field/index.ts
@@ -1,0 +1,1 @@
+export { LocationSelectField } from './LocationSelectField';
