+++ src/features/location/search-box/index.ts
@@ -1,0 +1,1 @@
+export { SearchBox } from './SearchBox';
