+++ src/features/location/search-tip/index.ts
@@ -1,0 +1,1 @@
+import { SearchTip } from './SearchTip';
