+++ src/pages/location-search/LocationSearch.style.ts
@@ -1,7 +1,5 @@
 import styled from '@emotion/styled';
 
-import Button from '#/components/button';
-
 export const Container = styled.div`
     height: 100%;
 `;
