+++ src/pages/location-search/index.ts
@@ -1,0 +1,1 @@
+export { LocationSearch } from './LocationSearch';
