+++ src/pages/templates/contents/Contents.tsx
@@ -1,13 +1,11 @@
 import { CSSProperties, PropsWithChildren } from 'react';
 
-import * as S from './index.style';
+import * as S from './Contents.style';
 
 export interface ContentsProps extends PropsWithChildren {
     style?: CSSProperties;
 }
 
-function Contents({ children, style }: ContentsProps) {
+export function Contents({ children, style }: ContentsProps) {
     return <S.Wrapper style={{ ...style }}>{children}</S.Wrapper>;
 }
-
-export default Contents;
