+++ src/pages/templates/contents/index.ts
@@ -1,0 +1,1 @@
+export { Contents } from './Contents';
